fix(project createBom): store multiple purls in property "purl_list"
My first try to store multiple purls separated by blanks (violating the
CycloneDX spec) and asking the user to fix the SBOM probably was a bad
idea from the beginning. Due to libary updates, the subsequent PURLs
were now encoded as part of the qualifiers which was hard to understand
and parse. So better keep the purl empty and store the list in a
separate property "purl_list".

Fixes #87