feat(bom map): check all releases for best matches
The previous code used to abort the mapping on the first good match. So
better or equal other matches could stay unnoticed. Also, the mapping
result may have depended on the order in which releases were found.

bom map --matchmode full-search allows to check all releases for
possible matches and thus assure that the mapping results are
deterministic and contain best matches.  To get the logic right, we also
need to do the checks in the order of MapResult priorities.