test(project): fix mocked JSON response for projects
SW360 REST API loves surprises in JSON encoding. It uses different ways
to describe linked releases in request and response bodies. The mocked
responses were seemingly copy&pasted from the requests, so they used the
wrong structure. This also fixes some copy&paste errors in the
release IDs so that `sw360:releases` and `linkedReleases` are
consistent.

While all of this had no effect on the test cases, it was confusing for
me when reading the code.