fix(bom map): use PURL from SW360 in output BOM
MapBom::update_bom_item() relied on an outdated (?) field
(`RepositoryId`) in the internal BOM data structure, which was not
populated for releases retrieved from SW360, so the PURL from the input
BOM was not replaced by SW360 data. Now rely on the result from
get_purl_from_match() which correctly looks up the purl in RepositoryId
as well as in `ExteralIds`.

Fixes #158